# perf(buffers): add batching to disk buffer reads