One last disk buffer thing now that we have more usable benchmarks :smile: 

This is a very simple, naive batching strategy for reads. The change is very isolated and shouldn't have any impact on correctness. It makes reading itself maybe a bit faster but more importantly seems to decrease contention/scheduling overhead by a meaningful amount when we're reading and writing simultaneously.

Benchmark results:
```
buffers/leveldb/writing time:   [329.06 ms 333.45 ms 338.33 ms]
                        thrpt:  [56.376 MiB/s 57.200 MiB/s 57.964 MiB/s]
                 change:
                        time:   [-0.1140% +1.3265% +2.6310%] (p = 0.09 > 0.05)
                        thrpt:  [-2.5636% -1.3092% +0.1141%]
                        No change in performance detected.

buffers/leveldb/reading time:   [576.98 ms 685.83 ms 737.68 ms]
                        thrpt:  [25.856 MiB/s 27.811 MiB/s 33.058 MiB/s]
                 change:
                        time:   [-24.644% -13.425% -1.4914%] (p = 0.07 > 0.05)
                        thrpt:  [+1.5140% +15.506% +32.704%]
                        No change in performance detected.

buffers/leveldb/both    time:   [1.0520 s 1.0563 s 1.0619 s]
                        thrpt:  [17.961 MiB/s 18.057 MiB/s 18.131 MiB/s]
                 change:
                        time:   [-9.2317% -8.4640% -7.8458%] (p = 0.00 < 0.05)
                        thrpt:  [+8.5138% +9.2467% +10.171%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low mild
```