@binarylogic 2020-04-28
What % are we talking from the previous numbers? Is this worth a ✨ highlight ✨ ? 🤩