@lukesteensen 2020-04-28
@binarylogic I think the line you're looking for is `thrpt:  [+8.5138% +9.2467% +10.171%]`, but I'd hesitate to say that will apply to all use cases. I'd be curious to see what the test-harness shows.